Adaptation 2G12-2
Déplacement 2G12-3 en 2G12-4
et scission de 2G12-2 en 2G12-2 et 2G12-3
