+++ src/csv/exosNonAmcNonInteractifs.csv
@@ -2,67 +2,67 @@
 
 6e,5e,4e,3e,2nde,1ere,Term,Reste 
 
-6C13,5A10,4C10-1,3A10,2F10-3,ref : 1E10,,CM020
-6C31-2,5A11-1,4C10-2,3A11-2,2F10-4,ref : 1E11,,CM021
-6C32,5A11,4C10-5,3A11-3,2G12-1,ref : 1E12,,PEA11-1
-6C35,5A12-1,4C10-6,3A12,2G12-2,ref : 1F10,,PEA11
-6D10,5A13,4C20-1,3A13,2G12-3,ref : 1N10,,P003
-6D11,5C11-1,4C20,3F1-act,2G22-1,ref : 1N11,,P004
-6D12,5C11-2,4C21-1,3F12,2G23-1,,,P005
-6G10-1,5C11,4C22-1,3F13,2G23-2,,,P006
-6G10-2,5C12-1,4C22-2,3F21,2G23,,,P007
-6G10,5C12,4C22,3G10-3,2G31-2,,,P008
-6G11,5G10-1,4C23,3G11,2G31-3,,,P009
-6G12-1,5G10-2,4C25-0,3G12,2G31-4,,,P010
-6G12,5G11-1,4C30-1,3G20-1,2G31-5,,,P011
-6G14,5G11-2,4C30-2,3G20-2,2G34-1,,,P012
-6G20-2,5G11-3,4C30-3,3G20,2G34-2,,,beta2F31
-6G20,5G12,4C30,3G22,2N11-1,,,beta2N60-X1
-6G21-1,5G13,4C32-2,3G23,2N11-2,,,beta2N60-X2
-6G21,5G20-1,4C32-3,3G30-1,2N14-1,,,beta3F23
-6G23-2,5G20,4C33-1,3G30-MG32,2N15-1,,,beta3G15
-6G23,5G21-1,4C33-3,3G31-1,2N15-2,,,beta3G41
-6G24-1,5G22-1,4G11,3G31-MG32,2N21,,,beta3s21
-6G24-2,5G22,4G15,3G32-2,2N32-1,,,beta4C31
-6G24-3,5G24-1,4G20-1,3G32-3,2N32-2,,,beta4G20-3
-6G25-3,5G24-2,4G20-5,3G32-4,2N32-3,,,beta4G20-4
-6G25,5G30-1,4G20MG32,3G32-5,2N32-4,,,beta6C33-1
-6G32-1,5G30,4G22,3G32,2N32-5,,,beta6test2
-6G42,5G31-1,4G30-1,3G40,2N32-6,,,beta6test2021
-6G43,5G31,4G31,3I1-1,2N41-7,,,betaAsymptotesObliques
-6I10,5G40-1,4G40-1,3I1,2N43-6,,,betaComplexes
-6I11,5G40-2,4G40-2,3L10-1,2N43-7,,,betaDivisionsDePolynomes
-6M11-2,5G40,4G40-MG32,3L10,2N52-1,,,betaEq1erDegreDansC
-6M22-1,5G41,4G41,3L13-1,2N52-2,,,betaEq2eDegAvecParam
-6M22-2,5G42,4G51,3L13-2,2N52-4,,,betaEqCarreDansC
-6M25,5G51,4G52,3L13,2N60-4,,,betaEqValAbs
-6M31-2,5L10-1,4I1,3L14-1,2N61-2,,,betaEquationsLog
-6N10-2,5L10-3,4L10-1,3L14,2N61-4,,,betaExo3d
-6N10-3,5L10-4,4L10,3L15,,,,betaExoSimpleMatthieu
-6N10-4,5L12-1,4L11,3S20,,,,betaModèle10_simple_question-reponse
-6N10-5,5L12,4L13-0,,,,,betaModèle11_paramétrable
-6N10,5L14-1,4L13-1,,,,,betaModèle20_plusieurs_types_de_questions
-6N11-3,5L14-3,4L14-0,,,,,betaModèle21_paramétrables
-6N11-4,5L14-4,4L14-1,,,,,betaModèle30_constructions_géométriques
-6N12,5L14-5,4L14-2,,,,,betaModèle31_paramétrables
-6N14,5L14-6,4L15-0,,,,,betaModèle40_tableau_proportionnalite
-6N20-1,5L15,4L15-1,,,,,betaModèle41_tableau_signes_variations
-6N22-1,5M10,4L16,,,,,betaProbaAouB
-6N22-2,5N11-1,4L20,,,,,betaProbabilites
-6N23-0,5N11-2,4P10-1,,,,,betaPuissances
-6N23-1,5N11-3,4P10-2,,,,,betaSys2x2CombLin
-6N23-2,5N11-4,4P10,,,,,betaTracerParabole
-6N23-3,5N14-2,4S10,,,,,betarotation3d
-6N23-4,5N14,4S11,,,,,betatrinome
-6N23,5N15,,,,,,moule_a_exo_mathalea
-6N24-1,5N20-0,,,,,,moule_a_exo_mathalea2d
-6N31-1,5P10,,,,,,c3C10-2
-6N31-2,5P12,,,,,,c3C10-4
-6N31-4,5R10-0,,,,,,c3C11
-6N31,5R10-1,,,,,,c3N10
-6N32,5R12-2,,,,,,c3N20
-6N33-0,5R12,,,,,,c3N22
-6N33-2,5R20-4,,,,,,c3N23
+6C13,5A10,4C10-1,3A10,2F10-3,ref : 1E10,,beta2F31
+6C31-2,5A11-1,4C10-2,3A11-2,2F10-4,ref : 1E11,,beta2N60-X1
+6C32,5A11,4C10-5,3A11-3,2G12-1,ref : 1E12,,beta2N60-X2
+6C35,5A12-1,4C10-6,3A12,2G12-2,ref : 1F10,,beta3F23
+6D10,5A13,4C20-1,3A13,2G12-3,ref : 1N10,,beta3G15
+6D11,5C11-1,4C20,3F1-act,2G12-4,ref : 1N11,,beta3G41
+6D12,5C11-2,4C21-1,3F12,2G22-1,,,beta3s21
+6G10-1,5C11,4C22-1,3F13,2G23-1,,,beta4C31
+6G10-2,5C12-1,4C22-2,3F21,2G23-2,,,beta4G20-3
+6G10,5C12,4C22,3G10-3,2G23,,,beta4G20-4
+6G11,5G10-1,4C23,3G11,2G31-2,,,beta6C33-1
+6G12-1,5G10-2,4C25-0,3G12,2G31-3,,,beta6test2
+6G12,5G11-1,4C30-1,3G20-1,2G31-4,,,beta6test2021
+6G14,5G11-2,4C30-2,3G20-2,2G31-5,,,betaAsymptotesObliques
+6G20-2,5G11-3,4C30-3,3G20,2G34-1,,,betaComplexes
+6G20,5G12,4C30,3G22,2G34-2,,,betaDivisionsDePolynomes
+6G21-1,5G13,4C32-2,3G23,2N11-1,,,betaEq1erDegreDansC
+6G21,5G20-1,4C32-3,3G30-1,2N11-2,,,betaEq2eDegAvecParam
+6G23-2,5G20,4C33-1,3G30-MG32,2N14-1,,,betaEqCarreDansC
+6G23,5G21-1,4C33-3,3G31-1,2N15-1,,,betaEquationsLog
+6G24-1,5G22-1,4G11,3G31-MG32,2N15-2,,,betaEqValAbs
+6G24-2,5G22,4G15,3G32-2,2N21,,,betaExo3d
+6G24-3,5G24-1,4G20-1,3G32-3,2N32-1,,,betaExoSimpleMatthieu
+6G25-3,5G24-2,4G20-5,3G32-4,2N32-2,,,betaModèle10_simple_question-reponse
+6G25,5G30-1,4G20MG32,3G32-5,2N32-3,,,betaModèle11_paramétrable
+6G32-1,5G30,4G22,3G32,2N32-4,,,betaModèle20_plusieurs_types_de_questions
+6G42,5G31-1,4G30-1,3G40,2N32-5,,,betaModèle21_paramétrables
+6G43,5G31,4G31,3I1-1,2N32-6,,,betaModèle30_constructions_géométriques
+6I10,5G40-1,4G40-1,3I1,2N41-7,,,betaModèle31_paramétrables
+6I11,5G40-2,4G40-2,3L10-1,2N43-6,,,betaModèle40_tableau_proportionnalite
+6M11-2,5G40,4G40-MG32,3L10,2N43-7,,,betaModèle41_tableau_signes_variations
+6M22-1,5G41,4G41,3L13-1,2N52-1,,,betaProbaAouB
+6M22-2,5G42,4G51,3L13-2,2N52-2,,,betaProbabilites
+6M25,5G51,4G52,3L13,2N52-4,,,betaPuissances
+6M31-2,5L10-1,4I1,3L14-1,2N60-4,,,betarotation3d
+6N10-2,5L10-3,4L10-1,3L14,2N61-2,,,betaSys2x2CombLin
+6N10-3,5L10-4,4L10,3L15,2N61-4,,,betaTracerParabole
+6N10-4,5L12-1,4L11,3S20,,,,betatrinome
+6N10-5,5L12,4L13-0,,,,,moule_a_exo_mathalea
+6N10,5L14-1,4L13-1,,,,,moule_a_exo_mathalea2d
+6N11-3,5L14-3,4L14-0,,,,,c3C10-2
+6N11-4,5L14-4,4L14-1,,,,,c3C10-4
+6N12,5L14-5,4L14-2,,,,,c3C11
+6N14,5L14-6,4L15-0,,,,,c3N10
+6N20-1,5L15,4L15-1,,,,,c3N20
+6N22-1,5M10,4L16,,,,,c3N22
+6N22-2,5N11-1,4L20,,,,,c3N23
+6N23-0,5N11-2,4P10-1,,,,,CM020
+6N23-1,5N11-3,4P10-2,,,,,CM021
+6N23-2,5N11-4,4P10,,,,,PEA11-1
+6N23-3,5N14-2,4S10,,,,,PEA11
+6N23-4,5N14,4S11,,,,,P003
+6N23,5N15,,,,,,P004
+6N24-1,5N20-0,,,,,,P005
+6N31-1,5P10,,,,,,P006
+6N31-2,5P12,,,,,,P007
+6N31-4,5R10-0,,,,,,P008
+6N31,5R10-1,,,,,,P009
+6N32,5R12-2,,,,,,P010
+6N33-0,5R12,,,,,,P011
+6N33-2,5R20-4,,,,,,P012
 6N33-3,5R20-5,,,,,,
 6N41-1,5R22-2,,,,,,
 6N43-4,5S10,,,,,,
